requirements.txt
# Dependencias principales del proyecto
# Framework web
flask>=3.0.0
Flask-CORS>=4.0.0

# ORM y Base de datos
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0

# Desarrollo
python-dotenv>=1.0.0
black>=23.0.0
flake8>=6.0.0
pylint>=3.0.0

# Utilidades
requests>=2.31.0
python-dateutil>=2.8.2
